Create GatherSortStudents.java
This script is the main test class for the Portfolio Project. It will allow the user to input names, addresses, and GPAs for as many Students as they like. It will then print out the list as they were input. Next it will sort the list by 'name and print out the list again. Additionally, it will save a version of the sorted list into a text file within the same directory as this file called: "Student_Log_File.txt".