Create PostfixCalculator.java
This java script will evaluate expressions that are in postfix notation. The script is also capable of evaluating the postfix expressions that are within a filepath separated by newlines.

The script is intelligent enough to notify the user when they entered an invalid postfix expression and when they have an expression that attempts to divide by zero.